last commit

+++ main/views.py
@@ -1,37 +1,53 @@
-from django.shortcuts import render, redirect
-from .forms import JssForm
-from .models import Jasoseol
+from django.shortcuts import render, redirect, get_object_or_404
+from .forms import JssForm, CommentForm
+from .models import Jasoseol, Comment
 from django.http import Http404
-
+from django.core.exceptions import PermissionDenied
+from django.contrib.auth.decorators import login_required
 
 def index(request):
     all_jss = Jasoseol.objects.all()
     return render(request, 'index.html', {'all_jss':all_jss})
 
+
+def my_index(request):
+    my_jss = Jasoseol.objects.filter(author=request.user)
+    return render(request, 'index.html', {'all_jss':my_jss})
+
+
+@login_required(login_url='/login/')
 def create(request):
-    print(request.user)
     if request.method == "POST":
         filled_form = JssForm(request.POST)
         if filled_form.is_valid():
-            temp_form = filled_form.save(commit=false)
+            temp_form = filled_form.save(commit=False)
             temp_form.author = request.user
-            filled_form.save()
+            temp_form.save()
             return redirect('index')
     jss_form = JssForm()
-    return render(request, 'create.html', {'jss_form' :jss_form})
+    return render(request, 'create.html', {'jss_form':jss_form})
 
+
+@login_required(login_url='/login/')
 def detail(request, jss_id):
-    try:
-        my_jss = Jasoseol.objects.get(pk=jss_id)
-    except:
-        raise Http404
+    # try: 
+    #     my_jss = Jasoseol.objects.get(pk=jss_id)
+    # except:
+    #     raise Http404
+    my_jss = get_object_or_404(Jasoseol, pk=jss_id)
+    comment_form = CommentForm()
 
-    return render(request, 'detail.html', {'my_jss' :my_jss})
+    return render(request, 'detail.html', {'my_jss':my_jss, 'commet_form':comment_form})
+
 
 def delete(request, jss_id):
     my_jss = Jasoseol.objects.get(pk=jss_id)
-    my_jss.delete()
-    return redirect('index')
+    if request.user == my_jss.author:
+        my_jss.delete()
+        return redirect('index')
+    
+    raise PermissionDenied
+    
 
 def update(request, jss_id):
     my_jss = Jasoseol.objects.get(pk=jss_id)
@@ -43,3 +59,26 @@ def update(request, jss_id):
             return redirect('index')
 
     return render(request, 'create.html', {'jss_form':jss_form})
+
+
+# comment 
+
+def create_comment(request, jss_id):
+    comment_form = CommentForm(request.POST)
+    if comment_form.is_valid():
+        temp_form = comment_form.save(commit=False)
+        temp_form.author = request.user
+        temp_form.jasoseol = Jasoseol.objects.get(pk=jss_id)
+        temp_form.save()
+        return redirect('detail', jss_id)
+
+
+def delete_comment(request, jss_id, comment_id):
+    my_comment = Comment.objects.get(pk=comment_id)
+    if request.user == my_comment.author:
+        my_comment.delete()
+        return redirect('detail', jss_id)
+
+    else:
+        raise PermissionDenied
+
